Now recognizes repeated random patterns
Previously the password 76441aaa76441 was not found to have a repeated pattern
in it. This is in contrast to say, something like apple2867apple where the
duplication of apple is recognized.

Before:
'76441' matches the pattern 'Random Characters with:Numbers'
	100 thousand passwords in the pattern
	49.74 percent of password strength
'aaa' matches the pattern 'Keyboard repeated character'
	1 thousand passwords in the pattern
	0.51 percent of password strength
'76441' matches the pattern 'Random Characters with:Numbers'
	100 thousand passwords in the pattern
	49.74 percent of password strength
Total passwords in all patterns: 10 trillion

After:
'76441' matches the pattern 'Random Characters with:Numbers'
    100 thousand passwords in the pattern
    98.98 percent of password strength
'aaa' matches the pattern 'Keyboard repeated character'
    1 thousand passwords in the pattern
    1.02 percent of password strength
'76441' matches the pattern 'Duplication of an earlier pattern: RANDOM_CHARACTERS'
    1 hundred passwords in the pattern
    0.00 percent of password strength
Total passwords in all finders: 103 million