Remove css height for finders at passfault.css (#60)
* Remove css height for finders at passfault.css

This makes the main page to be dozens of em's higher than needed.
Am I missing something or this property has no use?

* Force vertical scrollbar. Make pattern detail clickable
Change the way patterns are made visible: click instead of hover. This way, you can select text of the detail and read without getting the detail hiden.
Change cursor to pointer to indicate that pattern boxes are clickable.
Scroll to bottom when detail is showed.